+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 * `Next Release`_
 
   - Modify content negotiation to 406 when asked for an unknown character set
+  - Add :attr:`glinda.content.HandlerMixin.registered_content_types`
 
 * `0.0.3`_ (30 May 2015)